+++ module/nemd/traj.py
@@ -111,9 +111,7 @@ class Traj(list):
                     # FIXME: triclinic support
                     # The conventional units in the XTC file are nanometers and picoseconds.
                     box = pbc.Box.fromParams(*np.diag(box * 10))
-                    yield frame.Frame(xyz.astype(np.float64) * 10,
-                                      box=box,
-                                      step=step.item())
+                    yield frame.Frame(xyz * 10, box=box, step=step.item())
             return
         func = gzip.open if self.file.endswith('.gz') else open
         with func(self.file, 'rt') as fh:
